# Improve MLS IsSimple Algorithm