--- a/geom/attr_test.go
+++ b/geom/attr_test.go
@@ -172,6 +172,7 @@ func TestIsSimple(t *testing.T) {
 		{"POLYGON EMPTY", true},
 
 		{"MULTILINESTRING EMPTY", true},
+		{"MULTILINESTRING(EMPTY)", true},
 		{"MULTILINESTRING((0 0,1 0))", true},
 		{"MULTILINESTRING((0 0,1 0,0 1,0 0))", true},
 		{"MULTILINESTRING((0 0,1 1,2 2),(0 2,1 1,2 0))", false},