--- a/geom/attr_test.go
+++ b/geom/attr_test.go
@@ -178,6 +178,7 @@ func TestIsSimple(t *testing.T) {
 		{"MULTILINESTRING((0 0,1 1,2 2),(0 2,1 1,2 0))", false},
 		{"MULTILINESTRING((0 0,2 1,4 2),(4 2,2 3,0 4))", true},
 		{"MULTILINESTRING((0 0,2 0,4 0),(2 0,2 1))", false},
+		{"MULTILINESTRING((1 1,0 0,3 3,2 2),(1 1,2 2))", false},
 
 		// Cases for reproducing bugs.
 		{"MULTILINESTRING((0 0,0 1,1 1),(0 1,0 0,1 0))", false},
@@ -196,12 +197,12 @@ func TestIsSimple(t *testing.T) {
 		{"MULTIPOLYGON(((0 0,1 0,0 1,0 0)))", true},
 	} {
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			t.Logf("wkt: %s", tt.wkt)
 			got, defined := geomFromWKT(t, tt.wkt).IsSimple()
 			if !defined {
 				t.Fatal("not defined")
 			}
 			if got != tt.wantSimple {
-				t.Logf("wkt: %s", tt.wkt)
 				t.Errorf("got=%v want=%v", got, tt.wantSimple)
 			}
 		})