--- a/geom/attr_test.go
+++ b/geom/attr_test.go
@@ -186,11 +186,10 @@ func TestIsSimple(t *testing.T) {
 		{"MULTILINESTRING((0 0,3 0,3 3,0 3,0 0),(1 1,2 1,2 2,1 2,1 1))", true},
 		{"MULTILINESTRING((1 1,1 0,0 0),(1 1,0 1,0 0))", true},
 
-		// Cases for behaviour around duplicated lines. These cases are to
-		// match PostGIS and libgeos behaviour (the OGC spec is unclear about
-		// what the behaviour should be).
-		{"MULTILINESTRING((1 1,2 2),(1 1,2 2))", true},
-		{"MULTILINESTRING((1 1,2 2),(2 2,1 1))", true},
+		// Cases for behaviour around duplicated lines.
+		{"MULTILINESTRING((1 1,2 2),(1 1,2 2))", false},
+		{"MULTILINESTRING((1 1,2 2),(2 2,1 1))", false},
+		{"MULTILINESTRING((1 1,2 2,3 3),(1 1,2 2,3 3))", false},
 		{"MULTILINESTRING((1 1,2 2),(1 1,2 2,3 3))", false},
 		{"MULTILINESTRING((1 1,2 2),(2 2,1 1,3 1))", false},