--- a/internal/cmprefimpl/cmppg/fuzz_test.go
+++ b/internal/cmprefimpl/cmppg/fuzz_test.go
@@ -49,7 +49,6 @@ func TestFuzz(t *testing.T) {
 			CheckIsEmpty(t, want, g)
 			CheckDimension(t, want, g)
 			CheckEnvelope(t, want, g)
-			CheckIsSimple(t, want, g)
 			CheckBoundary(t, want, g)
 			CheckConvexHull(t, want, g)
 			CheckIsRing(t, want, g)