## Description

Eliminates O(n^2) worst case behaviour.

## Check List

Have you:

- Added unit tests? Yes.

- Add cmprefimpl tests? (if appropriate?) No (uses existing).

## Related Issue

- https://github.com/peterstace/simplefeatures/issues/166

## Benchmark Results
<details>
<summary>
Expand
</summary>


```
COMPARISON
name                                                      old time/op    new time/op    delta
MarshalWKB/polygon/n=10                                      222ns ± 1%     227ns ± 7%   +2.48%  (p=0.006 n=12+14)
MarshalWKB/polygon/n=100                                     443ns ± 6%     443ns ± 8%     ~     (p=0.547 n=13+15)
MarshalWKB/polygon/n=1000                                   2.42µs ± 2%    2.46µs ± 5%     ~     (p=0.247 n=14+15)
MarshalWKB/polygon/n=10000                                  21.5µs ±11%    21.1µs ± 2%     ~     (p=0.406 n=15+13)
UnmarshalWKB/polygon/n=10                                    399ns ± 2%     406ns ± 6%     ~     (p=0.125 n=14+15)
UnmarshalWKB/polygon/n=100                                   625ns ± 3%     633ns ± 6%     ~     (p=0.300 n=14+14)
UnmarshalWKB/polygon/n=1000                                 2.86µs ± 7%    2.84µs ± 3%     ~     (p=1.000 n=15+14)
UnmarshalWKB/polygon/n=10000                                28.6µs ± 6%    28.2µs ± 4%     ~     (p=0.254 n=15+13)
IntersectsLineStringWithLineString/n=10                     2.16µs ± 3%    2.16µs ± 3%     ~     (p=0.675 n=15+14)
IntersectsLineStringWithLineString/n=100                    29.7µs ± 4%    29.8µs ± 4%     ~     (p=0.928 n=13+15)
IntersectsLineStringWithLineString/n=1000                    306µs ± 4%     307µs ± 3%     ~     (p=0.943 n=14+13)
IntersectsLineStringWithLineString/n=10000                  4.00ms ±10%    4.00ms ± 5%     ~     (p=0.533 n=14+15)
IntersectsMultiPointWithMultiPoint/n=20                     1.81µs ± 2%    1.82µs ± 2%     ~     (p=0.096 n=14+14)
IntersectsMultiPointWithMultiPoint/n=200                    19.2µs ± 2%    19.2µs ± 1%     ~     (p=0.874 n=14+14)
IntersectsMultiPointWithMultiPoint/n=2000                    189µs ± 3%     190µs ± 2%     ~     (p=0.239 n=13+14)
IntersectsMultiPointWithMultiPoint/n=20000                  2.07ms ± 3%    2.09ms ± 5%     ~     (p=0.234 n=14+15)
LineStringIsSimpleZigZag/10                                 2.75µs ± 1%    2.77µs ± 3%     ~     (p=0.279 n=13+15)
LineStringIsSimpleZigZag/100                                45.1µs ± 3%    45.0µs ± 4%     ~     (p=0.769 n=14+14)
LineStringIsSimpleZigZag/1000                                501µs ± 2%     501µs ± 5%     ~     (p=0.400 n=14+15)
LineStringIsSimpleZigZag/10000                              7.14ms ± 6%    7.10ms ± 4%     ~     (p=0.813 n=15+14)
PolygonSingleRingValidation/n=10                            3.60µs ± 3%    3.61µs ± 4%     ~     (p=0.780 n=14+15)
PolygonSingleRingValidation/n=100                           51.3µs ± 6%    49.9µs ± 5%   -2.68%  (p=0.001 n=15+14)
PolygonSingleRingValidation/n=1000                           578µs ± 4%     579µs ± 3%     ~     (p=0.902 n=15+15)
PolygonSingleRingValidation/n=10000                         7.60ms ± 3%    7.63ms ± 4%     ~     (p=0.512 n=15+15)
PolygonMultipleRingsValidation/n=4                          10.8µs ± 4%    10.8µs ± 6%     ~     (p=0.813 n=14+15)
PolygonMultipleRingsValidation/n=36                         96.8µs ± 2%    96.7µs ± 4%     ~     (p=0.780 n=15+14)
PolygonMultipleRingsValidation/n=400                        1.29ms ± 2%    1.28ms ± 2%     ~     (p=0.067 n=15+15)
PolygonMultipleRingsValidation/n=4096                       16.2ms ± 5%    16.5ms ±10%     ~     (p=0.217 n=14+15)
PolygonZigZagRingsValidation/n=10                           18.0µs ± 1%    18.2µs ± 3%     ~     (p=0.067 n=12+14)
PolygonZigZagRingsValidation/n=100                           194µs ± 6%     191µs ± 3%     ~     (p=0.152 n=15+12)
PolygonZigZagRingsValidation/n=1000                         2.16ms ± 4%    2.15ms ± 8%     ~     (p=0.375 n=13+14)
PolygonZigZagRingsValidation/n=10000                        26.9ms ± 4%    26.6ms ± 5%     ~     (p=0.093 n=14+15)
PolygonAnnulusValidation/n=10                               5.44µs ± 5%    5.46µs ± 5%     ~     (p=0.691 n=15+14)
PolygonAnnulusValidation/n=100                              53.3µs ± 6%    52.5µs ± 2%     ~     (p=0.217 n=15+13)
PolygonAnnulusValidation/n=1000                              807µs ± 4%     805µs ± 3%     ~     (p=0.583 n=13+14)
PolygonAnnulusValidation/n=10000                            9.19ms ± 5%    9.19ms ± 3%     ~     (p=1.000 n=14+14)
MultipolygonValidation/n=1                                   421ns ± 4%     423ns ± 6%     ~     (p=0.609 n=15+13)
MultipolygonValidation/n=4                                  1.16µs ± 2%    1.17µs ± 7%     ~     (p=0.517 n=13+15)
MultipolygonValidation/n=16                                 7.02µs ± 3%    7.00µs ± 2%     ~     (p=0.847 n=14+15)
MultipolygonValidation/n=64                                 43.4µs ± 4%    43.1µs ± 3%     ~     (p=0.400 n=15+14)
MultipolygonValidation/n=256                                 258µs ± 2%     259µs ± 4%     ~     (p=0.756 n=13+14)
MultipolygonValidation/n=1024                               1.30ms ± 3%    1.30ms ± 4%     ~     (p=0.838 n=15+15)
MultiPolygonTwoCircles/n=10                                 5.27µs ± 3%    5.22µs ± 3%     ~     (p=0.311 n=15+13)
MultiPolygonTwoCircles/n=100                                60.2µs ± 3%    60.7µs ± 3%     ~     (p=0.102 n=15+14)
MultiPolygonTwoCircles/n=1000                                636µs ± 6%     626µs ± 3%     ~     (p=0.106 n=15+15)
MultiPolygonTwoCircles/n=10000                              9.88ms ± 4%   10.06ms ± 7%     ~     (p=0.089 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1                      6.84µs ± 6%    6.72µs ± 2%   -1.77%  (p=0.030 n=14+14)
MultiPolygonMultipleTouchingPoints/n=10                     55.9µs ± 4%    55.3µs ± 2%     ~     (p=0.201 n=15+14)
MultiPolygonMultipleTouchingPoints/n=100                     619µs ± 5%     613µs ± 1%     ~     (p=0.185 n=15+13)
MultiPolygonMultipleTouchingPoints/n=1000                   7.58ms ± 4%    7.57ms ± 4%     ~     (p=1.000 n=15+15)
Intersection/n=10                                            105µs ± 4%     103µs ± 1%   -2.33%  (p=0.000 n=15+14)
Intersection/n=100                                           637µs ± 4%     631µs ± 2%     ~     (p=0.158 n=15+14)
Intersection/n=1000                                         6.21ms ± 3%    6.19ms ± 7%     ~     (p=0.401 n=14+14)
Intersection/n=10000                                        68.3ms ± 4%    67.4ms ± 4%     ~     (p=0.102 n=14+15)
WKTParsing/point                                            2.10µs ± 7%    2.06µs ± 2%   -2.20%  (p=0.026 n=13+14)
DistancePolygonToPolygonOrdering/n=100_swap=false            105µs ± 2%     105µs ± 3%     ~     (p=0.595 n=15+15)
DistancePolygonToPolygonOrdering/n=100_swap=true             105µs ± 4%     105µs ± 4%     ~     (p=0.769 n=14+14)
DistancePolygonToPolygonOrdering/n=1000_swap=false          1.79ms ± 7%    1.77ms ± 2%     ~     (p=0.545 n=13+13)
DistancePolygonToPolygonOrdering/n=1000_swap=true           1.78ms ± 2%    1.77ms ± 2%     ~     (p=0.350 n=13+14)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false     9.40µs ± 2%    9.32µs ± 3%     ~     (p=0.171 n=12+14)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true      9.40µs ± 8%    9.49µs ± 6%     ~     (p=0.274 n=13+15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false     103µs ± 3%     102µs ± 4%     ~     (p=0.387 n=13+15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true      103µs ± 3%     102µs ± 4%     ~     (p=0.130 n=13+15)
MultiLineStringIsSimpleManyLineStrings/n=100                3.30ms ± 8%    0.08ms ± 2%  -97.71%  (p=0.000 n=14+14)
MultiLineStringIsSimpleManyLineStrings/n=1000                316ms ± 3%       1ms ± 4%  -99.74%  (p=0.000 n=13+14)

name                                                      old alloc/op   new alloc/op   delta
MarshalWKB/polygon/n=10                                       232B ± 0%      232B ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100                                    1.83kB ± 0%    1.83kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000                                   16.4kB ± 0%    16.4kB ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000                                   164kB ± 0%     164kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10                                     284B ± 0%      284B ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100                                  1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000                                 16.5kB ± 0%    16.5kB ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000                                 164kB ± 0%     164kB ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=10                     2.42kB ± 0%    2.42kB ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=100                    30.4kB ± 0%    30.4kB ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=1000                    205kB ± 0%     205kB ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=10000                  2.63MB ± 0%    2.63MB ± 0%     ~     (all equal)
IntersectsMultiPointWithMultiPoint/n=20                       325B ± 0%      324B ± 0%     ~     (p=1.000 n=15+15)
IntersectsMultiPointWithMultiPoint/n=200                    3.08kB ± 0%    3.08kB ± 0%     ~     (p=0.240 n=15+15)
IntersectsMultiPointWithMultiPoint/n=2000                   49.3kB ± 0%    49.3kB ± 0%     ~     (p=0.287 n=14+15)
IntersectsMultiPointWithMultiPoint/n=20000                   339kB ± 0%     339kB ± 0%     ~     (p=0.454 n=15+15)
LineStringIsSimpleZigZag/10                                 1.84kB ± 0%    1.84kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/100                                24.0kB ± 0%    24.0kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/1000                                139kB ± 0%     139kB ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/10000                              1.97MB ± 0%    1.97MB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=10                            2.19kB ± 0%    2.19kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=100                           24.3kB ± 0%    24.3kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=1000                           140kB ± 0%     140kB ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=10000                         1.97MB ± 0%    1.97MB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=4                          6.22kB ± 0%    6.22kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=36                         50.4kB ± 0%    50.4kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=400                         553kB ± 0%     553kB ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=4096                       5.67MB ± 0%    5.67MB ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=10                           9.38kB ± 0%    9.38kB ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=100                          87.8kB ± 0%    87.8kB ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=1000                          551kB ± 0%     551kB ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=10000                        7.24MB ± 0%    7.24MB ± 0%     ~     (all equal)
PolygonAnnulusValidation/n=10                               3.94kB ± 0%    3.94kB ± 0%     ~     (all equal)
PolygonAnnulusValidation/n=100                              28.3kB ± 0%    28.3kB ± 0%     ~     (all equal)
PolygonAnnulusValidation/n=1000                              379kB ± 0%     379kB ± 0%     ~     (all equal)
PolygonAnnulusValidation/n=10000                            3.89MB ± 0%    3.89MB ± 0%     ~     (all equal)
MultipolygonValidation/n=1                                    385B ± 0%      385B ± 0%     ~     (all equal)
MultipolygonValidation/n=4                                    676B ± 0%      676B ± 0%     ~     (all equal)
MultipolygonValidation/n=16                                 3.86kB ± 0%    3.86kB ± 0%     ~     (all equal)
MultipolygonValidation/n=64                                 15.4kB ± 0%    15.4kB ± 0%     ~     (all equal)
MultipolygonValidation/n=256                                63.1kB ± 0%    63.1kB ± 0%     ~     (all equal)
MultipolygonValidation/n=1024                                259kB ± 0%     259kB ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=10                                 4.99kB ± 0%    4.99kB ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=100                                55.0kB ± 0%    55.0kB ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=1000                                345kB ± 0%     345kB ± 0%     ~     (all equal)
MultiPolygonTwoCircles/n=10000                              4.60MB ± 0%    4.60MB ± 0%     ~     (p=0.696 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1                      3.95kB ± 0%    3.95kB ± 0%     ~     (all equal)
MultiPolygonMultipleTouchingPoints/n=10                     22.6kB ± 0%    22.6kB ± 0%     ~     (p=0.348 n=15+15)
MultiPolygonMultipleTouchingPoints/n=100                     172kB ± 0%     172kB ± 0%   -0.02%  (p=0.008 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1000                   2.04MB ± 0%    2.04MB ± 0%     ~     (p=0.563 n=15+13)
Intersection/n=10                                           27.2kB ± 0%    27.2kB ± 0%   -0.03%  (p=0.026 n=15+15)
Intersection/n=100                                           140kB ± 0%     140kB ± 0%     ~     (p=0.940 n=14+15)
Intersection/n=1000                                         1.69MB ± 0%    1.69MB ± 0%     ~     (p=0.345 n=15+15)
Intersection/n=10000                                        17.8MB ± 0%    17.8MB ± 0%     ~     (p=0.461 n=15+15)
WKTParsing/point                                            1.93kB ± 0%    1.93kB ± 0%     ~     (all equal)
DistancePolygonToPolygonOrdering/n=100_swap=false           40.7kB ± 0%    40.7kB ± 0%     ~     (p=1.000 n=15+15)
DistancePolygonToPolygonOrdering/n=100_swap=true            40.7kB ± 0%    40.7kB ± 0%     ~     (p=0.133 n=13+15)
DistancePolygonToPolygonOrdering/n=1000_swap=false           369kB ± 0%     369kB ± 0%     ~     (p=0.115 n=15+14)
DistancePolygonToPolygonOrdering/n=1000_swap=true            369kB ± 0%     369kB ± 0%     ~     (p=0.976 n=15+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false     5.52kB ± 0%    5.52kB ± 0%     ~     (all equal)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true      5.52kB ± 0%    5.52kB ± 0%     ~     (all equal)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false    60.1kB ± 0%    60.1kB ± 0%     ~     (p=0.196 n=15+15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true     60.1kB ± 0%    60.1kB ± 0%     ~     (p=1.000 n=14+15)
MultiLineStringIsSimpleManyLineStrings/n=100                2.17MB ± 0%    0.06MB ± 0%  -97.28%  (p=0.000 n=15+15)
MultiLineStringIsSimpleManyLineStrings/n=1000                216MB ± 0%       0MB ± 0%  -99.77%  (p=0.000 n=15+15)

name                                                      old allocs/op  new allocs/op  delta
MarshalWKB/polygon/n=10                                       6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=100                                      6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=1000                                     6.00 ± 0%      6.00 ± 0%     ~     (all equal)
MarshalWKB/polygon/n=10000                                    6.00 ± 0%      6.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10                                     7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=100                                    7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=1000                                   7.00 ± 0%      7.00 ± 0%     ~     (all equal)
UnmarshalWKB/polygon/n=10000                                  7.00 ± 0%      7.00 ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=10                       9.00 ± 0%      9.00 ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=100                      73.0 ± 0%      73.0 ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=1000                      345 ± 0%       345 ± 0%     ~     (all equal)
IntersectsLineStringWithLineString/n=10000                   5.46k ± 0%     5.46k ± 0%     ~     (all equal)
IntersectsMultiPointWithMultiPoint/n=20                       1.00 ± 0%      1.00 ± 0%     ~     (all equal)
IntersectsMultiPointWithMultiPoint/n=200                      7.00 ± 0%      7.00 ± 0%     ~     (all equal)
IntersectsMultiPointWithMultiPoint/n=2000                     6.00 ± 0%      6.00 ± 0%     ~     (all equal)
IntersectsMultiPointWithMultiPoint/n=20000                    11.0 ± 0%      11.0 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/10                                   7.00 ± 0%      7.00 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/100                                  71.0 ± 0%      71.0 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/1000                                  343 ± 0%       343 ± 0%     ~     (all equal)
LineStringIsSimpleZigZag/10000                               5.46k ± 0%     5.46k ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=10                              9.00 ± 0%      9.00 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=100                             73.0 ± 0%      73.0 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=1000                             345 ± 0%       345 ± 0%     ~     (all equal)
PolygonSingleRingValidation/n=10000                          5.46k ± 0%     5.46k ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=4                            39.0 ± 0%      39.0 ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=36                            313 ± 0%       313 ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=400                         3.43k ± 0%     3.43k ± 0%     ~     (all equal)
PolygonMultipleRingsValidation/n=4096                        35.1k ± 0%     35.1k ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=10                             38.0 ± 0%      38.0 ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=100                             230 ± 0%       230 ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=1000                          1.05k ± 0%     1.05k ± 0%     ~     (all equal)
PolygonZigZagRingsValidation/n=10000                         16.4k ± 0%     16.4k ± 0%     ~     (all equal)
PolygonAnnulusValidation/n=10                                 19.0 ± 0%      19.0 ± 0%     ~     (all equal)
PolygonAnnulusValidation/n=100                                73.0 ± 0%      73.0 ± 0%     ~     (all equal)
PolygonAnnulusValidation/n=1000                              1.00k ± 0%     1.00k ± 0%     ~     (all equal)
PolygonAnnulusValidation/n=10000                             10.2k ± 0%     10.2k ± 0%     ~     (all equal)
MultipolygonValidation/n=1                                    5.00 ± 0%      5.00 ± 0%     ~     (all equal)
MultipolygonValidation/n=4                                    8.00 ± 0%      8.00 ± 0%     ~     (all equal)
MultipolygonValidation/n=16                                   27.0 ± 0%      27.0 ± 0%     ~     (all equal)
MultipolygonValidation/n=64                                   99.0 ± 0%      99.0 ± 0%     ~     (all equal)
MultipolygonValidation/n=256                                   392 ± 0%       392 ± 0%     ~     (all equal)
…